Add a GitHub workflow to prepare a release pull request
This automates the first half of the current release process. The second
half will be implemented by another workflow.

The reason why it can't all be done in a single workflow is that it's useful
to let developers inspect what'll go into the release before actually
publishing it, and to apply any last-minute fixes, if necessary. It also
allows CI to complete for the release PR.

To support the new workflow, add a `--set` option to `update_version.py`
that sets the version to a custom value.