Add a GitHub workflow for finalizing a release
This automates the second half of the release process (the first being
automated by `prepare-release.yml`). After this workflow completes, the only
action that should remain for the releaser to do is to merge the
`dev-release-*` pull request. We can't do that as part of this workflow,
because CI has to finish first, and it seems pointless to create another
workflow just to merge 1 PR.

Make the release notes extraction process more sophisticated to work around
GitHub's frustrating handling of line breaks in PR and release
descriptions.